+++ src/main/java/wooteco/subway/admin/controller/AdminController.java
@@ -1,0 +1,43 @@
+package wooteco.subway.admin.controller;
+
+import org.springframework.http.MediaType;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import wooteco.subway.admin.repository.StationRepository;
+import wooteco.subway.admin.service.LineService;
+
+@RequestMapping("/admin")
+@Controller
+public class AdminController {
+    private LineService lineService;
+    private StationRepository stationRepository;
+
+    public AdminController(LineService lineService, StationRepository stationRepository) {
+        this.lineService = lineService;
+        this.stationRepository = stationRepository;
+    }
+
+    @GetMapping(value = "/", produces = MediaType.TEXT_HTML_VALUE)
+    public String index(Model model) {
+        return "admin/index";
+    }
+
+    @GetMapping(value = "/stations", produces = MediaType.TEXT_HTML_VALUE)
+    public String stationPage(Model model) {
+        model.addAttribute("stations", stationRepository.findAll());
+        return "admin/admin-station";
+    }
+
+    @GetMapping(value = "/lines", produces = MediaType.TEXT_HTML_VALUE)
+    public String linePage(Model model) {
+        model.addAttribute("lines", lineService.showLines());
+        return "admin/admin-line";
+    }
+
+    @GetMapping(value = "/edges", produces = MediaType.TEXT_HTML_VALUE)
+    public String edgePage(Model model) {
+        return "admin/admin-edge";
+    }
+}
