+++ src/main/java/wooteco/subway/admin/controller/PageController.java
@@ -22,23 +22,6 @@ public class PageController {
         return "service/index";
     }
 
-    @GetMapping(value = "/stations", produces = MediaType.TEXT_HTML_VALUE)
-    public String stationPage(Model model) {
-        model.addAttribute("stations", stationRepository.findAll());
-        return "admin/admin-station";
-    }
-
-    @GetMapping(value = "/lines", produces = MediaType.TEXT_HTML_VALUE)
-    public String linePage(Model model) {
-        model.addAttribute("lines", lineService.showLines());
-        return "admin/admin-line";
-    }
-
-    @GetMapping(value = "/edges", produces = MediaType.TEXT_HTML_VALUE)
-    public String edgePage(Model model) {
-        return "admin/admin-edge";
-    }
-
     @GetMapping(value = "/search", produces = MediaType.TEXT_HTML_VALUE)
     public String searchPage(Model model) {
         return "service/search";
